feat!: pre-command input
The Action now allows a set of arbitrary commands to be executed prior to invoking Terragrunt,
which allows for customization of the container image (such as adding SSH keys or other Git configuration).

BREAKING CHANGE: This is the first major release for this Action. (No breaking changes actually included;
this message is purely to trigger the major version bump.)